+++ SOSAuto_BACKEND/src/main/java/com/sosauto_backend/model/Entity/Véhicule.java
@@ -13,7 +13,7 @@ public class Véhicule {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int véhiculeid;
+    private Long véhiculeid;
     private String marque;
     private String modele;
     private int année;
